Limits are updated on new d/w/m even without connection
The limits are reset by Timekpr every new day, week or month.
The webfrontend until now showed only the current values from the database.
If there hasn't been a connection to the device for a long time, the limits will be out of date.
In this case, the frontend tries to show the limits "that should be active" right now.

Meaning:
Daily limit will be reset after midnight to maximum, even taking into account that weekly/monthly limit might be smaller.
Weekly limit will be reset on Monday.
Monthly limit will be reset on the first day of the month.
Playtime limit will be reset after midnight.